+++ src/main/java/com/github/touhoumaidaffection/TouhouMaidAffection.java
@@ -14,6 +14,7 @@ import com.github.touhoumaidaffection.handler.LapPillowPoseConfigHandler;
 import com.github.touhoumaidaffection.handler.MorningKissVoiceConfigHandler;
 import com.github.touhoumaidaffection.handler.RescueActionConfigHandler;
 import com.github.touhoumaidaffection.handler.RescueVoiceConfigHandler;
+import com.github.touhoumaidaffection.handler.TmaAiStatusRequestHandler;
 import com.github.touhoumaidaffection.handler.VoicePreviewRequestHandler;
 import com.github.touhoumaidaffection.handler.TmaSettingsRequestHandler;
 import com.github.touhoumaidaffection.network.BondActivateAbilityPayload;
@@ -33,6 +34,9 @@ import com.github.touhoumaidaffection.network.MorningKissVoiceConfigPayload;
 import com.github.touhoumaidaffection.network.MorningKissVoicePlayPayload;
 import com.github.touhoumaidaffection.network.RescueActionConfigPayload;
 import com.github.touhoumaidaffection.network.RescueVoiceConfigPayload;
+import com.github.touhoumaidaffection.network.TmaAiCacheClearPayload;
+import com.github.touhoumaidaffection.network.TmaAiStatusPayload;
+import com.github.touhoumaidaffection.network.TmaAiStatusRequestPayload;
 import com.github.touhoumaidaffection.network.TmaSettingsRequestPayload;
 import com.github.touhoumaidaffection.network.TmaSettingsStatePayload;
 import com.github.touhoumaidaffection.network.VoicePreviewDataPackPlayPayload;
@@ -114,7 +118,12 @@ public class TouhouMaidAffection {
         id = registerMessage(id, LapPillowStartPayload.class, LapPillowStartPayload.STREAM_CODEC, LapPillowHandler::handleStart);
         id = registerMessage(id, LapPillowExitPayload.class, LapPillowExitPayload.STREAM_CODEC, LapPillowHandler::handleExit);
         id = registerMessage(id, LapPillowPoseConfigPayload.class, LapPillowPoseConfigPayload.STREAM_CODEC, LapPillowPoseConfigHandler::handle);
-        registerMessage(id, LapPillowAngleLockPayload.class, LapPillowAngleLockPayload.STREAM_CODEC, LapPillowAngleLockHandler::handle);
+        id = registerMessage(id, LapPillowAngleLockPayload.class, LapPillowAngleLockPayload.STREAM_CODEC, LapPillowAngleLockHandler::handle);
+        id = registerMessage(id, TmaSettingsRequestPayload.class, TmaSettingsRequestPayload.STREAM_CODEC, TmaSettingsRequestHandler::handle);
+        id = registerMessage(id, TmaSettingsStatePayload.class, TmaSettingsStatePayload.STREAM_CODEC, BondClientPayloadHandler::handleSettingsState);
+        id = registerMessage(id, TmaAiStatusRequestPayload.class, TmaAiStatusRequestPayload.STREAM_CODEC, TmaAiStatusRequestHandler::handleStatusRequest);
+        id = registerMessage(id, TmaAiStatusPayload.class, TmaAiStatusPayload.STREAM_CODEC, BondClientPayloadHandler::handleAiStatus);
+        registerMessage(id, TmaAiCacheClearPayload.class, TmaAiCacheClearPayload.STREAM_CODEC, TmaAiStatusRequestHandler::handleClear);
     }
 
     private <T> int registerMessage(
@@ -134,8 +143,6 @@ public class TouhouMaidAffection {
                     context.setPacketHandled(true);
                 }
         );
-        id = registerMessage(id, TmaSettingsRequestPayload.class, TmaSettingsRequestPayload.STREAM_CODEC, TmaSettingsRequestHandler::handle);
-        id = registerMessage(id, TmaSettingsStatePayload.class, TmaSettingsStatePayload.STREAM_CODEC, BondClientPayloadHandler::handleSettingsState);
         return id + 1;
     }
 }